python-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyMuPDF==1.23.8
Pillow==10.1.0
pytesseract==0.3.10
langchain==0.1.0
langchain-groq==0.0.1
langchain-community==0.0.13
sentence-transformers==2.7.0
faiss-cpu==1.8.0
python-dotenv==1.0.0